add exe, move msk token provider to its own file
