+++ Gemfile
@@ -12,3 +12,4 @@ gem "rubocop-minitest", "0.34.5"
 gem "rubocop-packaging", "0.5.2"
 gem "rubocop-performance", "1.20.2"
 gem "rubocop-rake", "0.6.0"
+gem "thor", "~> 1.3"
