+++ exe/signer
@@ -1,0 +1,4 @@
+#!/usr/bin/env ruby
+
+require "aws/msk/iam/sasl/signer"
+Aws::Msk::Iam::Sasl::Signer::CLI.start(ARGV)
